+++ src/aks-preview/azext_aks_preview/custom.py
@@ -618,7 +618,7 @@ def aks_create(
     # azure service mesh
     enable_azure_service_mesh=None,
     revision=None,
-    # azure monitor profile
+    # azure monitor profile - metrics
     enable_azuremonitormetrics=False,
     enable_azure_monitor_metrics=False,
     azure_monitor_workspace_resource_id=None,
@@ -626,6 +626,8 @@ def aks_create(
     ksm_metric_annotations_allow_list=None,
     grafana_resource_id=None,
     enable_windows_recording_rules=False,
+    # azure monitor profile - app monitoring
+    enable_azure_monitor_app_monitoring=False,
     # metrics profile
     enable_cost_analysis=False,
     # AI toolchain operator
@@ -808,6 +810,9 @@ def aks_update(
     enable_windows_recording_rules=False,
     disable_azuremonitormetrics=False,
     disable_azure_monitor_metrics=False,
+    # azure monitor profile - app monitoring
+    enable_azure_monitor_app_monitoring=False,
+    disable_azure_monitor_app_monitoring=False,
     enable_vpa=False,
     disable_vpa=False,
     enable_addon_autoscaling=False,